# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tangent Blog OpenApi Demo Documentation",
        "description": "Tangent Blog Swagger OpenApi description",
        "contact": {
            "email": "sipho411@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://tangent_blog.test",
            "description": "Demo Tangent Blog API Server"
        }
    ],
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get all posts for authenticated user",
                "description": "Get all posts for authenticated user",
                "operationId": "postList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/post/store": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Add a new post as Auth user",
                "description": "Add a new post as Auth user",
                "operationId": "postStore",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/post/{id}/show": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get a post by Id",
                "description": "Get a post by id.",
                "operationId": "postShow",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/post/{id}update": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Update a post as Auth user",
                "description": "Update a post as Auth user",
                "operationId": "postUpdate",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/post/{id}/destroy": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Delete a post",
                "description": "Delete a post",
                "operationId": "postDestroy",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/post/{keyword}/search": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Search posts",
                "description": "Search posts",
                "operationId": "postSearch",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/post/getPostComments": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Add a new post as Auth user",
                "description": "Add a new post as Auth user",
                "operationId": "getPostComments",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "list users",
                "description": "list users",
                "operationId": "users",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get user details",
                "description": "Get user details",
                "operationId": "userShow",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App/Http/Resources/Api/V1/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
...
}